requirements.txt
python-socketio==5.11.1
eventlet==0.35.2
